refactor FB package naming
