+++ packages/ui-components/src/components/Wallet/Send.test.tsx
@@ -4,8 +4,8 @@ import React from 'react';
 import * as fireBlocksActions from '../../actions/fireBlocksActions';
 import * as pav3Actions from '../../actions/pav3Actions';
 import {CustodyState} from '../../lib';
-import * as fireBlocksState from '../../lib/fireBlocks/storage/state';
 import type {GetEstimateTransactionResponse} from '../../lib/types/fireBlocks';
+import * as fireBlocksState from '../../lib/wallet/storage/state';
 import {VALID_ETH_ADDRESS} from '../../tests/common';
 import {mockAccountAsset, mockWallets} from '../../tests/mocks/wallet';
 import {customRender} from '../../tests/test-utils';
