--- packages/ui-components/src/lib/fireBlocks/index.ts
@@ -1,2 +1,0 @@
-export * from './reactSigner';
-export * from './storage';
