+++ packages/ui-components/src/lib/index.ts
@@ -1,7 +1,6 @@
 export * from './domain';
 export * from './error';
 export * from './fetchApi';
-export * from './fireBlocks';
 export * from './formOpenseaLink';
 export * from './formSocialMediaLink';
 export * from './i18n';
