+++ packages/ui-components/src/lib/wallet/index.ts
@@ -1,3 +1,5 @@
 export * from './login';
+export * from './reactSigner';
 export * from './signer';
+export * from './storage';
 export * from './tokens';
